requirements.txt
pydantic

scikit-learn

pandas

numpy

pyyaml

pydantic-settings

pytest